Avoid taking ownership if not needed, reorganize modules structure
